Add microtonal (fractional) MIDI pitch (#36)
Migrate MidiNote.pitch from int to double so a note can land between
equal-tempered semitones, carrying microtonal / just-intonation / arbitrary
tunings through the whole transform chain. Whole-number pitches behave exactly
as before.

- New ScaleTuning (cents-per-degree); ScaleConformanceTransform snaps in cents,
  so non-12-TET scales work (.diatonic reproduces the old behaviour).
- Inversion keeps the fractional reflection; Transpose/Splitting preserve the
  fractional offset; SpectralMapping maps to fractional targets; voice-routing
  scale-degree match rounds to nearest semitone.
- Piano-roll yForPitch renders fractional pitches between lanes; authoring
  still snaps to whole semitones.
- MidiGateway gains pitchBend; EngineMidiController gains an opt-in microtonal
  flag that voices fractional pitch as nearest-semitone + per-channel bend
  (GM +-2 range). Off by default, it just rounds.
- SMF reader -> double; writer rounds to nearest semitone (a .mid can't carry
  cents).

+++ integration_test/midi_loop_transform_test.dart
@@ -60,11 +60,11 @@ void main() {
     final looped = const SmfReader().read(fakeIo.savedBytes!).notes;
 
     expect(looped.length, unlooped.length * 2);
-    final unloopedCounts = <int, int>{};
+    final unloopedCounts = <double, int>{};
     for (final n in unlooped) {
       unloopedCounts[n.pitch] = (unloopedCounts[n.pitch] ?? 0) + 1;
     }
-    final loopedCounts = <int, int>{};
+    final loopedCounts = <double, int>{};
     for (final n in looped) {
       loopedCounts[n.pitch] = (loopedCounts[n.pitch] ?? 0) + 1;
     }
